backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
motor>=3.3.0
python-dotenv>=1.0.0
transformers>=4.30.0,<4.40.0
torch>=1.13.0,<2.1.0
torchvision>=0.14.0,<0.16.0
Pillow>=9.0.0,<10.1.0
requests>=2.31.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pydantic>=2.5.2 